moving nodes into their own folder
