+++ test/unit/objectExplorerUtils.test.ts
@@ -7,7 +7,7 @@ import * as path from "path";
 import { ObjectExplorerUtils } from "../../src/objectExplorer/objectExplorerUtils";
 import { expect, assert } from "chai";
 import * as Constants from "../../src/constants/constants";
-import { TreeNodeInfo } from "../../src/objectExplorer/treeNodeInfo";
+import { TreeNodeInfo } from "../../src/objectExplorer/nodes/treeNodeInfo";
 import { ConnectionProfile } from "../../src/models/connectionProfile";
 import { ObjectMetadata } from "vscode-mssql";
